Move flavours/glitch/util/api to flavours/glitch/api
