+++ app/javascript/flavours/glitch/components/admin/Dimension.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import api from 'flavours/glitch/util/api';
+import api from 'flavours/glitch/api';
 import { FormattedNumber } from 'react-intl';
 import { roundTo10 } from 'flavours/glitch/util/numbers';
 import Skeleton from 'flavours/glitch/components/skeleton';
